MAINT: Move common code to TyphonDesignerMixin
