+++ typhon/display.py
@@ -3,12 +3,12 @@ import logging
 import os.path
 
 from pydm import Display
-from pydm.widgets.base import PyDMWidget
 from qtpy.QtCore import Property, Slot, Q_ENUMS
 from qtpy.QtWidgets import QHBoxLayout
 
 from .utils import ui_dir, TyphonBase, clear_layout
-from .plugins import HappiChannel
+from .widgets import TyphonDesignerMixin
+
 
 logger = logging.getLogger(__name__)
 
@@ -29,7 +29,7 @@ class TemplateTypes:
         return Enum('TemplateEnum', entries)
 
 
-class TyphonDisplay(TyphonBase, PyDMWidget, TemplateTypes):
+class TyphonDisplay(TyphonBase, TyphonDesignerMixin, TemplateTypes):
     """
     Main Panel display for a signal Ophyd Device
 
@@ -58,13 +58,6 @@ class TyphonDisplay(TyphonBase, PyDMWidget, TemplateTypes):
 
     parent: QWidget, optional
     """
-    # Unused properties that we don't want visible in designer
-    alarmSensitiveBorder = Property(bool, designable=False)
-    alarmSensitiveContent = Property(bool, designable=False)
-    precisionFromPV = Property(bool, designable=False)
-    precision = Property(int, designable=False)
-    showUnits = Property(bool, designable=False)
-
     # Template types and defaults
     Q_ENUMS(TemplateTypes)
     TemplateEnum = TemplateTypes.to_enum()  # For convenience
@@ -74,7 +67,6 @@ class TyphonDisplay(TyphonBase, PyDMWidget, TemplateTypes):
 
     def __init__(self,  parent=None, **kwargs):
         # Intialize background variable
-        self._channel = None
         self._use_template = ''
         self._use_default = False
         self._last_macros = dict()
@@ -87,31 +79,6 @@ class TyphonDisplay(TyphonBase, PyDMWidget, TemplateTypes):
         self.layout().setContentsMargins(0, 0, 0, 0)
         # Load template
         self.load_template()
-
-    @Property(str)
-    def channel(self):
-        """The channel address to use for this widget"""
-        if self._channel:
-            return str(self._channel)
-        return None
-
-    @channel.setter
-    def channel(self, value):
-        if self._channel != value:
-            # Remove old connection
-            if self._channels:
-                self._channels.clear()
-                for channel in self._channels:
-                    if hasattr(channel, 'disconnect'):
-                        channel.disconnect()
-            # Load new channel
-            self._channel = str(value)
-            channel = HappiChannel(address=self._channel,
-                                   tx_slot=self._tx)
-            self._channels = [channel]
-            # Connect the channel to the HappiPlugin
-            if hasattr(channel, 'connect'):
-                channel.connect()
 
     @property
     def current_template(self):
